Move schemas to separate file
Move the pydantic classes/schemas used to validate the requests
and responses formats to `chatbot_api/schemas.py`. This practice
is necessary for a modularized, easy to maintain project. The
schemas are necessary to validate the requests received by the
API, as well as to reformat the data sent back to the client.
